CORRECTIONS I HAVE PROPOSED IN ll. 11 (two) and l. 14
PNYU 1a (P.Corn. inv. II 37) is from Karanis and is dated to A.D. 318-320. There are some corrections on the text after its restoration. The image is available in http://quod.lib.umich.edu/a/apis/x-1254/c2_37r.tif. In l. 11: between δέκα πέντε, there is trace of a letter and a space for 2-3 letters and we could assume δέκα κ[αὶ] πέντε; cf. P.Cair.Isid. 77, 6 (Karanis; A.D. 320). Then, instead of ιγιρ, where there is no trema on the first ι, probably we should read Ἠ̣ρακλ̣  ̣  ̣  ̣.  Also, in l. 13: instead of λοιπω read λοιπῶν.