SOME CORRECTION BASED ON THE PRIGINAL AND THE PHOTO ONLINE in ll. 3, 10 and 16 (added)
P. Fay. 243 (P.Corn. inv. MSS. A 101. XV) was described in the first edition, and then L.C. Youtie published the verso in ZPE 50 (1983), pp. 51-55 (Tafel IIa) (= SB XVI 13001). The recto was still described in the footnote 1 of this edition and it contains “the names of the individuals, who may have been praktors credited with collections on behalf of villages” and was published as SB XVI 13000 (L.C. Youtie brieflich). The image is available in http://quod.lib.umich.edu/a/apis/x-3432/c0_xvr.tif. Only note l. 3, where Youtie read  Ὀφελλι̣[. The round left part of alpha is clear after the restoration, so Ὀφελλᾶ[ς is rather probable. In l. 10 Youtie read Ευ̣  ̣ ο  ̣  ̣  ̣ως̣. There are some stray blots of ink at that part of the papyrus, which make the reading difficult. I propose to read Εὐδ̣α̣ίμων   ̣[. After the certain Ευ the letter δ is not made as in the line below. Also, α is made as a single oblique stroke ligatured with the following ι. Finally there is one more line that can be seen after the restoration of the papyrus, which starts with an α: l. 16 α[  . . . ] . .[.

 - SOME CORRECTION BASED ON THE PRIGINAL AND THE PHOTO ONLINE in ll. 3, 10 and 16 (added)
 - SOME CORRECTION BASED ON THE PRIGINAL AND THE PHOTO ONLINE in ll. 3, 10 and 16 (added)

P. Fay. 243 (P.Corn. inv. MSS. A 101. XV) was described in the first edition, and then L.C. Youtie published the verso in ZPE 50 (1983), pp. 51-55 (Tafel IIa) (= SB XVI 13001). The recto was still described in the footnote 1 of this edition and it contains “the names of the individuals, who may have been praktors credited with collections on behalf of villages” and was published as SB XVI 13000 (L.C. Youtie brieflich). The image is available in http://quod.lib.umich.edu/a/apis/x-3432/c0_xvr.tif. Only note l. 3, where Youtie read  Ὀφελλι̣[. The round left part of alpha is clear after the restoration, so Ὀφελλᾶ[ς is rather probable. In l. 10 Youtie read Ευ̣  ̣ ο  ̣  ̣  ̣ως̣. There are some stray blots of ink at that part of the papyrus, which make the reading difficult. I propose to read Εὐδ̣α̣ίμων   ̣[. After the certain Ευ the letter δ is not made as in the line below. Also, α is made as a single oblique stroke ligatured with the following ι. Finally there is one more line that can be seen after the restoration of the papyrus, which starts with an α: l. 16 α[  . . . ] . .[.
 - some new readings in l. 3, 9, 10, 16

Signed-off-by: Joshua D. Sosin <joshuad.sosin@gmail.com>
Signed-off-by: James Cowey <james.cowey@urz.uni-heidelberg.de>
Signed-off-by: Rodney Ast <simoeis@gmail.com>
Signed-off-by: Paul Heilporn <paul.heilporn@unistra.fr>
Signed-off-by: Thomas Kruse <thomas.kruse@oeaw.ac.at>
Signed-off-by: Cisca Hoogendijk <ciscahoogendijk@live.nl>
Signed-off-by: Graham Claytor <gclaytor@gmail.com>
Signed-off-by: Mark Depauw <mark.depauw@arts.kuleuven.be>
Signed-off-by: Lajos Berkes <berkes.lajos@googlemail.com>
Signed-off-by: Alain Delattre <adelattr@ulb.ac.be>
Signed-off-by: Jenny Cromwell <jenniferacromwell@gmail.com>
Signed-off-by: Tonia Sarri <toniasarri@gmail.com>
Signed-off-by: Matias Buchholz <matias.buchholz@helsinki.fi>
Signed-off-by: Christian-Jürgen Gruber <morunculus@gmail.com>
Signed-off-by: Hélène Cuvigny <h.cuvigny@gmail.com>
Signed-off-by: W. Graham Claytor <gclaytor@umich.edu>
Signed-off-by: Arthur Verhoogt <verhoogt@umich.edu>