Text from ed. pr.
Since I published the text, I dare to correct some minor errors:
l. 3-6, 9: I corrected καμήλλ(ια) into καμήλ(ια), since double λ  is written because of the "pluralische Kürzung" as stated in note to l. 2.
l. 8: I typed the reading given in the commentary ad locum which I intended to write in the edition: σπεκου]λ̣λά̣τ̣ορ(ος) instead of σπεκου]λ̣ά̣τ̣ορ(ος)

I was not able to add the brackets correctly in l. 8.

 - Created from SoSOL template
 - Text from ed. pr.

Since I published the text, I dare to correct some minor errors:
l. 3-6, 9: I corrected καμήλλ(ια) into καμήλ(ια), since double λ  is written because of the "pluralische Kürzung" as stated in note to l. 2.
l. 8: I typed the reading given in the commentary ad locum which I intended to write in the edition: σπεκου]λ̣λά̣τ̣ορ(ος) instead of σπεκου]λ̣ά̣τ̣ορ(ος)

I was not able to add the brackets correctly in l. 8.

Signed-off-by: Joshua D. Sosin <joshuad.sosin@gmail.com>
Signed-off-by: James Cowey <james.cowey@urz.uni-heidelberg.de>
Signed-off-by: Rodney Ast <simoeis@gmail.com>
Signed-off-by: Paul Heilporn <paul.heilporn@unistra.fr>
Signed-off-by: Thomas Kruse <thomas.kruse@oeaw.ac.at>
Signed-off-by: Cisca Hoogendijk <ciscahoogendijk@live.nl>
Signed-off-by: Graham Claytor <gclaytor@gmail.com>
Signed-off-by: Mark Depauw <mark.depauw@arts.kuleuven.be>
Signed-off-by: Lajos Berkes <berkes.lajos@googlemail.com>
Signed-off-by: Alain Delattre <adelattr@ulb.ac.be>
Signed-off-by: Jenny Cromwell <jenniferacromwell@gmail.com>
Signed-off-by: Tonia Sarri <toniasarri@gmail.com>
Signed-off-by: Matias Buchholz <matias.buchholz@helsinki.fi>
Signed-off-by: Christian-Jürgen Gruber <morunculus@gmail.com>
Signed-off-by: Hélène Cuvigny <h.cuvigny@gmail.com>
Signed-off-by: W. Graham Claytor <gclaytor@umich.edu>
Signed-off-by: Arthur Verhoogt <verhoogt@umich.edu>