I propose a minor correction:
ἐπικ(ειμένῳ) -->(ἐπικ(ειμένου)

The construction epi+gen. is usual to introduce the name of the pagarch in this period, see Federico Morelli in notes to CPR XXII 18, 11. This fits also the first genitive  υἱο(ῦ) which was corrected in the ed. pr. (APF 31 (1985) 17) to dative.

Besides, I contracted line 2 and 3, since there is no line 3 in the ed. pr and SB.

 - I propose a minor correction:

ἐπικ(ειμένῳ) -->(ἐπικ(ειμένου)

The construction epi+gen. is usual to introduce the name of the pagarch in this period, see Federico Morelli in notes to CPR XXII 18, 11. This fits also the first genitive  υἱο(ῦ) which was corrected in the ed. pr. (APF 31 (1985) 17) to dative.

Besides, I contracted line 2 and 3, since there is no line 3 in the ed. pr and SB.

Signed-off-by: Joshua D. Sosin <joshuad.sosin@gmail.com>
Signed-off-by: James Cowey <james.cowey@urz.uni-heidelberg.de>
Signed-off-by: Rodney Ast <simoeis@gmail.com>
Signed-off-by: Paul Heilporn <paul.heilporn@unistra.fr>
Signed-off-by: Thomas Kruse <thomas.kruse@oeaw.ac.at>
Signed-off-by: Cisca Hoogendijk <ciscahoogendijk@live.nl>
Signed-off-by: Graham Claytor <gclaytor@gmail.com>
Signed-off-by: Mark Depauw <mark.depauw@arts.kuleuven.be>
Signed-off-by: Lajos Berkes <berkes.lajos@googlemail.com>
Signed-off-by: Alain Delattre <adelattr@ulb.ac.be>
Signed-off-by: Jenny Cromwell <jenniferacromwell@gmail.com>
Signed-off-by: Tonia Sarri <toniasarri@gmail.com>
Signed-off-by: Matias Buchholz <matias.buchholz@helsinki.fi>
Signed-off-by: Christian-Jürgen Gruber <morunculus@gmail.com>
Signed-off-by: Hélène Cuvigny <h.cuvigny@gmail.com>
Signed-off-by: W. Graham Claytor <gclaytor@umich.edu>
Signed-off-by: Arthur Verhoogt <verhoogt@umich.edu>