Update revisionDesc
 - Vote - Accept-Straight-to-Finalization - R:
only one line:
1. <:((ὑπὲρ)) (τιμ(ῆς))=N. Gonis, ZPE 150 (2004) 193|ed|(δι(ὰ)) (|Τιμ|):> εἰδῶν <:λόγῳ|reg|λωγῶν:> <:<:τῶν|reg|δων:> <:κ[.?]|alt|κ[αράβων]:>=N. Gonis, ZPE 150 (2004) 193|ed|Δανιῆ[λ]:> [.?]
and add a commentary: The same text is repeated in 9 lines, but not always in complete form, due to the damaged status of the papyrus. The text given is a reconstruction of the maximum legible text possible.

V:

add a commentary: The same text is repeated in 9 lines, but not always in complete form, due to the damaged status of the papyrus. The text given is a reconstruction of the maximum legible text possibl


 (Lajos Berkes)
 - Finalized - something went wrong with the commentary to the recto. I'll add commentary to this line later. (Lajos Berkes)