text copied and edited.
l.1: I added para in brackets, but the reading at the end of the line is incorrect as I know from unpublished corrections of N. Gonis. (In fact, para is there and other names should be read.)
l.2: The second correction stems from F. Morelli and not Th. Kurse as it is mentioned in a footnote in APF.

 - Created from SoSOL template
 - text copied and edited.

l.1: I added para in brackets, but the reading at the end of the line is incorrect as I know from unpublished corrections of N. Gonis. (In fact, para is there and other names should be read.)
l.2: The second correction stems from F. Morelli and not Th. Kurse as it is mentioned in a footnote in APF.

Signed-off-by: Joshua D. Sosin <joshuad.sosin@gmail.com>
Signed-off-by: James Cowey <james.cowey@urz.uni-heidelberg.de>
Signed-off-by: Rodney Ast <simoeis@gmail.com>
Signed-off-by: Paul Heilporn <paul.heilporn@unistra.fr>
Signed-off-by: Thomas Kruse <thomas.kruse@oeaw.ac.at>
Signed-off-by: Cisca Hoogendijk <ciscahoogendijk@live.nl>
Signed-off-by: Graham Claytor <gclaytor@gmail.com>
Signed-off-by: Mark Depauw <mark.depauw@arts.kuleuven.be>
Signed-off-by: Lajos Berkes <berkes.lajos@googlemail.com>
Signed-off-by: Alain Delattre <adelattr@ulb.ac.be>
Signed-off-by: Jenny Cromwell <jenniferacromwell@gmail.com>
Signed-off-by: Tonia Sarri <toniasarri@gmail.com>
Signed-off-by: Matias Buchholz <matias.buchholz@helsinki.fi>
Signed-off-by: Christian-Jürgen Gruber <morunculus@gmail.com>
Signed-off-by: Hélène Cuvigny <h.cuvigny@gmail.com>
Signed-off-by: W. Graham Claytor <gclaytor@umich.edu>
Signed-off-by: Arthur Verhoogt <verhoogt@umich.edu>