Rescued a broken file. Original submission note is:
typed in from the edition. l1 suffered from severe typos, I tried to reconstruct it on the base of the commentary. I didn't enter the supralines at the beginning of ll. 3-6: I don't think this is necessary. l7: κγ doesn't show as a number in the preview.

 - Created from SoSOL template
 - Rescued a broken file. Original submission note is:

typed in from the edition. l1 suffered from severe typos, I tried to reconstruct it on the base of the commentary. I didn't enter the supralines at the beginning of ll. 3-6: I don't think this is necessary. l7: κγ doesn't show as a number in the preview.

Signed-off-by: Joshua D. Sosin <joshuad.sosin@gmail.com>
Signed-off-by: James Cowey <james.cowey@urz.uni-heidelberg.de>
Signed-off-by: Rodney Ast <simoeis@gmail.com>
Signed-off-by: Paul Heilporn <paul.heilporn@unistra.fr>
Signed-off-by: Thomas Kruse <thomas.kruse@oeaw.ac.at>
Signed-off-by: Cisca Hoogendijk <ciscahoogendijk@live.nl>
Signed-off-by: Graham Claytor <gclaytor@gmail.com>
Signed-off-by: Mark Depauw <mark.depauw@arts.kuleuven.be>
Signed-off-by: Lajos Berkes <berkes.lajos@googlemail.com>
Signed-off-by: Alain Delattre <adelattr@ulb.ac.be>
Signed-off-by: Jenny Cromwell <jenniferacromwell@gmail.com>
Signed-off-by: Tonia Sarri <toniasarri@gmail.com>
Signed-off-by: Matias Buchholz <matias.buchholz@helsinki.fi>
Signed-off-by: Christian-Jürgen Gruber <morunculus@gmail.com>
Signed-off-by: Hélène Cuvigny <h.cuvigny@gmail.com>
Signed-off-by: W. Graham Claytor <gclaytor@umich.edu>
Signed-off-by: Arthur Verhoogt <verhoogt@umich.edu>