added corrections from BL 2.2 with some of my own from autopsy.
The name Φαήρμ(ιος), attested only here, has been tentatively read as a form of the common name Φάησις.  The ed.'s rho does not seem likely, since elsewhere the tail of rho comes straight down or has a slight hook (although cf. Τραιανοῦ, l. 14), and the ligature to the next letter seems to come from high on the letter, which is unlikely for rho, but fits sigma (cf. Σαμβαθίῳ l. 4) or epsilon. On the other hand, fitting σεω in after Φαη seems difficult, and it seems somewhat hazardous to suggest Φαή<σ>ε̣ω̣(ς) vel sim. when there's so much abrasion.  

The name Phaesis, however read, is perhaps guaranteed by the Tax Rolls, which have a Phaesis s. of Pekmeis (alias Heuremon) and Sambathion (P.Mich. IV.2, p. 244), who's also probably found in BGU 15 2498 (143/4?).  This is likely the son of the Pekmeis in this document, who's acting as kyrios for his wife Sambathion.

The other changes are minor; the app. crit. seems a bit overloaded when adding diaereseis, but they weren't in the original edition.

If accepted, I'll add some short commentary. 

 - added corrections from BL 2.2 with some of my own from autopsy.

The name Φαήρμ(ιος), attested only here, has been tentatively read as a form of the common name Φάησις.  The ed.'s rho does not seem likely, since elsewhere the tail of rho comes straight down or has a slight hook (although cf. Τραιανοῦ, l. 14), and the ligature to the next letter seems to come from high on the letter, which is unlikely for rho, but fits sigma (cf. Σαμβαθίῳ l. 4) or epsilon. On the other hand, fitting σεω in after Φαη seems difficult, and it seems somewhat hazardous to suggest Φαή<σ>ε̣ω̣(ς) vel sim. when there's so much abrasion.

The name Phaesis, however read, is perhaps guaranteed by the Tax Rolls, which have a Phaesis s. of Pekmeis (alias Heuremon) and Sambathion (P.Mich. IV.2, p. 244), who's also probably found in BGU 15 2498 (143/4?).  This is likely the son of the Pekmeis in this document, who's acting as kyrios for his wife Sambathion.

The other changes are minor; the app. crit. seems a bit overloaded when adding diaereseis, but they weren't in the original edition.

If accepted, I'll add some short commentary.
 - Resubmission.  I have addressed the editors' comments by adding commentary to ll. 2 and 6 and have corrected Leiden+ errors.  I was unable to add links in the commentary b/c of a bug (?): whenever I selected the volume, then volume number, all the fields would go grey and I couldn't selected an individual text.  I tried this both through the DDB and HGV fields.

Signed-off-by: Joshua D. Sosin <joshuad.sosin@gmail.com>
Signed-off-by: James Cowey <james.cowey@urz.uni-heidelberg.de>
Signed-off-by: Rodney Ast <simoeis@gmail.com>
Signed-off-by: Paul Heilporn <paul.heilporn@unistra.fr>
Signed-off-by: Thomas Kruse <thomas.kruse@oeaw.ac.at>
Signed-off-by: Cisca Hoogendijk <ciscahoogendijk@live.nl>
Signed-off-by: Graham Claytor <gclaytor@gmail.com>
Signed-off-by: Mark Depauw <mark.depauw@arts.kuleuven.be>
Signed-off-by: Lajos Berkes <berkes.lajos@googlemail.com>
Signed-off-by: Georg Schmelz <georg.schmelz@gmx.de>
Signed-off-by: Alain Delattre <adelattr@ulb.ac.be>
Signed-off-by: Jenny Cromwell <jenniferacromwell@gmail.com>
Signed-off-by: Tonia Sarri <toniasarri@gmail.com>
Signed-off-by: Christian-Jürgen Gruber <morunculus@gmail.com>
Signed-off-by: Hélène Cuvigny <h.cuvigny@gmail.com>
Signed-off-by: W. Graham Claytor <gclaytor@umich.edu>
Signed-off-by: Arthur Verhoogt <verhoogt@umich.edu>
Signed-off-by: Esther Garel <esthergarel@gmail.com>