Update revisionDesc
 - Vote - AcceptText - Checked Clarysse's readings on the original. The only remaining problem that there really seems to be a phi in the middle of Clarysse's Σεβε̣ννύ(). (Arthur Verhoogt)
 - Vote - DEFER - I agree with all of Clarysse's improvements, but, like Verhoogt, I see a phi in lin13. I see sigma with epsilon written above it; then beta, epsilon, phi; then two rearing nus, the first one quite close to the phi and the second one with an upsilon written above it. Simplest thing to do in my view is to email Clarysse seeking clarification in the light of the APIS image: http://quod.lib.umich.edu/a/apis/x-3535/6980V.TIF ; BTW, what are we to make of what looks to be an ancient diacritical in lin13 ὀφε'ίλον(τας) ? (Joshua D. Sosin)
 - Vote - AcceptText - See notes. Fine. (James Cowey)
 - Finalized - Ready. Clarysse notes, of the phi, "I see it too now, but what can you do with that? Put it between { } brackets? Could I be completely wrong and should we read two words after all: Se( ) befnny( ) or something with nê(sos) at the end? I must have thought this trace was just a shadow on the papyrus but it clearly is not." After multiple emails with Clarysse, J. Sosin and he propose encoding as follows: <:<:<:(Σεβεννύ(του))(?)|reg|(Σεβε̣{φ}ννύ(του)):>|alt|(|Σε|) (|βεφννθ|):>=PN W. Clarysse (digital image)||ed||(Σεβε̣ννύ(του))=W. Clarysse, APF 48 (2002) 101| (|σε|) (|.1φ̣ν̣ο̣ν.1|):> 
I emailed Clarysse as follows: James recently added your improvements to ZPE 105 (1995) 237-243 [=SB 22 15767]. It is in voting now. Arthur and I note that at line 13 the reading seems to be σεβεφννυ. In other words, there is an phi in the middle of Σεβε̣ννύ(του). This seems pretty clear in the APIS image: http://quod.lib.umich.edu/a/apis/x-3535/6980V.TIF Could you please have a look and explain/reconsider in the light of the image, and get back to us? 
I emailed Clarysse as follows: James recently added your improvements to ZPE 105 (1995) 237-243 [=SB 22 15767]. It is in voting now. Arthur and I note that at line 13 the reading seems to be σεβεφννυ. In other words, there is an phi in the middle of Σεβε̣ννύ(του). This seems pretty clear in the APIS image: http://quod.lib.umich.edu/a/apis/x-3535/6980V.TIF Could you please have a look and explain/reconsider in the light of the image, and get back to us?  (James Cowey)