Update revisionDesc
 - Vote - AcceptText - l1: => (Γεωργι(ο  ))
delete ll.15-19 and add commentary: The same text is repeated in 5 lines (Ἰουλίου being abbreviated as Ἰουλί(ου)), but not always in complete form, due to the damaged status of the papyrus. The text given is a reconstruction of the maximum legible text possible.
delete ll. 21-25: and add a commentary: The same text is repeated in 5 lines, but not always in complete form, due to the damaged status of the papyrus. The text given is a reconstruction of the maximum legible text possible (Lajos Berkes)
 - Vote - AcceptText - Accept. Lin14-25 I might fuss over a couple of the udnerdots, but do not think it worth the trouble. This fuller accounting than the ed gives is in my view useful. (Joshua D. Sosin)
 - Finalized - ready (Lajos Berkes)