Update revisionDesc
 - Vote - AcceptText - lin2 =>There are no lines lost.
lin3 => I would rather have the names in column since it is an account. It is difficult to number the lines though, because the text doesn't seem to be fully edited. Next voter ? (Esther Garel)
 - Vote - AcceptText - The text is a descriptum from P.Ryl.copt. 232.
use 00. ~| |~cop for the Coptic words.
l1: =>  (|ε̣τ̣πρ|)(|θ|) =>  <:(εἰ(σ)πραχθ(έντων))(?)=N. Gonis, ZPE 131 (2000) 152|ed|(|ε̣τ̣πρ|)(|θ|):> (underdot ei) (Lajos Berkes)
 - Finalized - ready (Lajos Berkes)