Update revisionDesc
 - Vote - AcceptText - L. 1: On this dark image I think I see the usual abbreviation αδελ(φῷ) with a rather wide lambda written over the epsilon. What was read as the tail of the supposed phi is actually the top part of the epsilon. No underdots needed, too.
L. 6: The editor has λέγ⟨ε⟩εις τῶι   (l. ποιεῖς λέγων αὐτῶι), which, however, is quite an intrusive alteration. The DDBDP did not represent the real reading of the editor, someone seems to have it silently changed into <λ>έγε εις τῶι. This again, does not seem to be compatible with the photo. The first letter is rather lambda than epsilon. There is no way of finding out what the editor meant by adding this epsilon, but I think Josh is right in thinking it must have been λέγ⟨ε⟩ις. This could indeed be corrected (again) silently.
Note that the line break before l. 6 is missing, and there are some superfluous spaces in l. 5. (Cisca Hoogendijk)
 - Vote - AcceptText - Ln1: I am willing to accept (ἀδελ̣(φῷ)), though I find it hard to make out all the letters. The beginning of ln 6 is difficult and I am not entirely persuaded by the readings discussed thus far.  It makes sense however to add what the editor presumably meant, i.e. λέγ⟨ε⟩ις. (Rodney Ast)
 - Finalized - Ready. I found it very difficult to verify these readings from the online image, so I contacted M. Gerhardt in Berlin and asked him to check the original. This is what he reported: "I have read the suggestions for BGU XVI 2650 and I agree with you all: in line 1 I can clearly read adel(fw) with a wide lambda written over the epsilon (ok, it is a bit fainted); no underdots needed. In line 6 I can clearly read leg<e>is. I think both are typos in Brashear's editio princeps."
  (James Cowey)