Update revisionDesc
 - Vote - AcceptText - lin10 Σεργ̣ί̣[ο]υ - the gamma is certainly better than an eta, which has a long high first vertical stroke. The iota descends quite markedly and more than sometimes, but cf. lin15 εἰς where in ligature the iota is indeed long. So, yes, I reckon the new reading is distinctly better and there is little that makes it unlikely.
lin18 τετραξε(στιαῖα) the small alpha attached to the xi, which descends markedly, is clear. The raised epsilon is also quite clear.
lin29 (καὶ) this is certainly comparable to SB VI 9589,16. P.Oxy. LVIII 3933,33 is not a good parallel as far as I can see. In the Oxy papyrus the καὶ is clearly visible in full, I think.
Remove the rasied letter mark up throughout. Otherwise fine. (James Cowey)
 - Vote - AcceptText - line 10: Yes, gamma is preferable to eta following rho. Σεργ̣ί̣[ο]υ is thus better than Σερή̣ρ̣ο̣υ̣.  Line 18: good reading.  I am willing to accept the proposal γραφὲ(ν) (καὶ) ἐπερ(ωτη)θ(εὶς), although I agree with Cowey that P.Oxy. 58.3933.13 is not an clear parallel.  I've not been able to verify the symbol in SB 6.9589 due to problems with the website in Vienna. (Rodney Ast)
 - Finalized - Ready. (James Cowey)