Update revisionDesc
 - Submit - Added the Greek text. Some notes and questions:
Column i, line 1.: according to the printed edition there is a smaller and raised gamma. I am not sure if my interpretation is correct. The editor has not given any comment on it. Could the gamma mark an abbreviation? Or, could it perhaps be a numeral (= 83)?
Column ii, lines 6, 7, 9, 10 and 11: there are marks on top of the first letters. Do they mark numerals? The editor seems to consider these (marked?) letters as numerals. Should those marks be written here as well and if so, how? (Katariina Kankaanpaa)
 - Vote - Accept-Straight-to-Finalization - Fine (James Cowey)
 - Finalized - Ready. (James Cowey)