Update 62564.xml
I enter P.Oxy. 26.2442 + 75.5039 (TM 62564), proofread against the print edition.
The texts has various points that need to be re-considered: see the tickets #114, #121, #165
Some symbols might be discussed in the commentary (in a few cases I wrote *mark*).
Sometimes it was not easy to distinguish between text written above the line and marginalia.