Update revisionDesc
 - Submit - Line 3: the symbol is for δεκανίας, not ἑκατονταρχίας, and Kronios is therefore a cameldriver and not a soldier, which makes sense.  In line 2, we presume that the symbol for δεκανίας is in the lacuna on the left--this would explain why Kronios is said to be from the _same_ dekania.  Line 4: we've dotted the stigma and, in the comm., suggested the alternative reading of gamma.  Line 5: ἐλαβ(  ) αν  ̣  ̣(  ) of the ed.pr. is discarded in favor of the equally obscure αλαβανω(  ), which we conjecture is a personal name. Line 6: instead of ἐλαβ̣(  ) plus traces, read [Πε]τ̣εασ̣μ̣(ήφιος) [κ]ε̣[ρ](άμια) ι.  We've also added an introduction and commentary notes.  An infrared photo is available upon request. (Rodney Ast)
 - Vote - AcceptText - I agree for the replacement of (ἑκατονταρχίαϲ) by (δεκανίαϲ) at lines 2 and 3.
I am more reserved for changes at lines 5-6. At the end of line 6, κερα( ) ι is alright (I consider that the curved stroke above represents the α (the same is true for lines 2 and 3), but it needs a lot of imagination to see τεασμ, even with dots… As for ἐλαβ,  ε was practically invisible, and the proposed  α instead is as well. After λαβ there is space for one letter, then what Rodney and Roger read α (think of χ?) then ν. I would at least dot all the letters except λαβ and ν   (Hélène Cuvigny)
 - Vote - Return-Straight-to-Sender - Back again for further updates. (James Cowey)
 - Vote - Accept-Straight-to-Finalization - Looks good. (Julia Lougovaya)
 - Finalized - Ready (Rodney Ast)