Update revisionDesc
 - Commit - Addition of σε after ἀξιῶ δὲ from an examination of a scan supplied by the host institution (Patrick James)
 - Submit - I was suspicious of the syntaxt of ἀξιῶ δὲ ἀντιφωνεῖν and expected there to be an accusative 'you'. A nominative and infinitive 'I think it right that I reply' would be at odds with [μ]οι.

Depending on the scenario understood (as to which party should reply to which more frequently), the writer could claim that messages will come from him and that he thinks it right for him to reply to the addressee more frequently. If so, [σ]οι could be read, and no accusative would be necessary.

However, an examination in a graduate seminar of a scan supplied by Barry Hopkins of McCormick Theological Seminar revealed that the pronoun σε had been omitted both by Grenfell and Hunt and by Witkowski. On the basis of the scan, Sarah Baribeau of Trinity Western University pointed out the omission in the transcription and I confirmed it. (Patrick James)
 - Vote - Return-Straight-to-Sender - Patrick: This is a good catch and a good suggestion. Could you please (1) email me (joshua.sosin@duke.edu) and James Cowey (james.cowey@urz.uni-heidelberg.de) the image on basis of which you add the new reading and (2) simply re-submit the text? We can then verify and, if warranted as seems likely, accept. Also, you may wish (though we do not require it) to add line breaks. (Joshua D. Sosin)
 - Submit - Addition of σε after ἀξιῶ δὲ from an examination of a scan supplied by the host institution. 

I was suspicious of the syntaxt of ἀξιῶ δὲ ἀντιφωνεῖν and expected there to be an accusative 'you'. A nominative and infinitive 'I think it right that I reply' would be at odds with [μ]οι. Depending on the scenario understood (as to which party should reply to which more frequently), the writer could claim that messages will come from him and that he thinks it right for him to reply to the addressee more frequently. If so, [σ]οι could be read, and no accusative would be necessary. However, an examination in a graduate seminar of a scan supplied by Barry Hopkins of McCormick Theological Seminar revealed that the pronoun σε had been omitted both by Grenfell and Hunt and by Witkowski. On the basis of the scan, Sarah Baribeau of Trinity Western University pointed out the omission in the transcription and I confirmed it. (Patrick James)
 - Vote - AcceptText - Nice catch.  Please attribute the emendation thus: <:σε=PN P. James (with S. Baribeau; from a photo)|ed|:> . Also, P. James has, on the basis of the photo (confirmed by me), noted the following line divisions. Line 2. ζητῶ, 3. ἐν, 4. καλαὶ, 5. [ἐ]μοῦ, 6. [μ]οι, 7.- τας, 8. ἐπιμέλου, 9. ἔρρω(σο). These should be added at finalization. James plans to make further improvements to the text in the coming weeks. (Rodney Ast)
 - Vote - AcceptText - Fine. See previous votes. Put in the line breaks on finalisation. (James Cowey)
 - Finalized - Ready (Rodney Ast)