Update revisionDesc
 - Commit - P.Corn. 38 is a list of names, probably liturgists (demosioi?). The last two lines are described as “fragments of 2 lines”. The first of them is actually crossed out with horizontal strokes. The second can be read as Πανεχώτης Βελλ[ ]. The name Πανεχώτης appears in l. 9, where the editors read Πανεκώτης and comment in the introduction that the name is new. However, even in l. 9, the letter χ is clear and we should read Πανεχώτης. The image is available in http://quod.lib.umich.edu/a/apis/x-1206/c1_105r.tif (nikos litinas)
 - Submit - JUST SOME CORRECTIONS I HAVE PROPOSED (in l. 9 and in two last lines not transcribed in the ed. princ., now ll. 14 and 15) THEY ARE BASED ON THE ORIGINAL AND THE PHOTO ONLINE:
P.Corn. 38 is a list of names, probably liturgists (demosioi?). The last two lines are described as “fragments of 2 lines”. The first of them is actually crossed out with horizontal strokes. The second can be read as Πανεχώτης Βελλ[ ]. The name Πανεχώτης appears in l. 9, where the editors read Πανεκώτης and comment in the introduction that the name is new. However, even in l. 9, the letter χ is clear and we should read Πανεχώτης. The image is available in http://quod.lib.umich.edu/a/apis/x-1206/c1_105r.tif (nikos litinas)
 - Vote - Return-Straight-to-Sender - Sending back to let you look at this again, as we decided. (James Cowey)
 - Vote - AcceptText - Accept with some changes. Lin9 : I am not 100% persuaded of the new reading Π̣α̣ν̣ε̣κώτης  ==> Π̣α̣ν̣ε̣χώτης ; see the kappa in Κοπρέ[ου]ς (lin3), which is not all that different. Kappa in lin9 (as ed pr) seems plausible to me; so, I'd urge Π̣α̣ν̣ε̣χώτης either go in the app or else be accepted with (?) in the lem. Lin15 : I cannot see horizontal deletion strokes, but only a smudge; cannot tell from the digital image (http://quod.lib.umich.edu/cgi/i/image/api/image/apis/X-1206/C1_105R.TIF/full/large/0/native.jpg) whether that smudge is erasure or transfer. Lin16 : I'd be inclined to dot some of the letters: Π̣ανεχώ̣τ̣ης̣ Βελλ[.?] .  (Joshua D. Sosin)
 - Vote - AcceptText - Accept. Line 16: Π̣ανεχώ̣τ̣ης Βελλ[ is a sound reading (I follow Josh Sosin on the fact that some letters should be dotted). Line 15 looks indeed like (voluntary) erasure, leaving only a smudge. As for the name in line 9, it should be noted that this would be, at the time being, the only attestation of the spelling Πανεκώτης (instead of the usual Πανεχώτης as in line 16); it's true that, at first sight, the fifth letter looks like a κ, but one should note tha (Paul Heilporn)
 - General - By mistake, I hit voting before finishing my explanation. Here it is:
Accept. Line 16: Π̣ανεχω̣τ̣ης Βελλ[ is a sound reading (I follow Josh Sosin on the fact that some letters should be dotted). Line 15 looks indeed like (voluntary) erasure, leaving only a smudge. As for the name in line 9, it should be noted that this would be, at the time being, the only attestation of the spelling Πανεκωτης (instead of the usual Πανεχωτης as in line 16; there is a Πανεκατις in P.Bon. 11); it's true that, at first sight, the fifth letter looks like a κ, but one should note that the only other κ (l. 3, as pointed out by Josh Sosin) starts much higher (top left). Therefore, I think Nikos Litinas is right in recognizing this as a χ in ligature with the previous ε (the stroke coming from the ε goes to the upper right corner of χ; from there, without interruption, he makes the diagonal stroke to the lower left corner, then (again without interruption?) he goes to the upper left corner and makes the diagonal stroke to the lower right corner and from there to the following ω.
By the way doesn't this look like a list of liturgists or δημόσιοι (one son of Sarapion being named as ---φύλαξ and the new Panekhotes son of Belles being named as arkhephodos)? Compare e.g. P.Petaus 67 (185); P.Oxy. LXI 4128 (346). (Paul Heilporn)
 - Finalized - Ready. (James Cowey)