Update revisionDesc
 - Submit - Würzberg workshop; Submitted according to the edition.
The columns which don't have accent are the columns of the transliterated Latin in the Greek alphabets.
The biblio num is unknown and I encoded the deleted letters by upper or lower dots or a cancel stroke using the double square brackets. 
In the notes, it is mentioned that there is space for nine more lines at the foot of the 1st column below line 35, but this is not mentioned in the text itself.  (Fatma E. Hamouda)
 - Vote - Accept-Straight-to-Finalization - I'd record cancellation by dots, i.e. col. 1, l. 5, "the ε is deleted by dots above and below." Col. 3, line 5: record diacritical over iota; col. 3, line 40: take out regularization since readings in the Latin column are not regularized anywhere else. Otherwise fine. (Julia Lougovaya-Ast)
 - Finalized - Ready (Rodney Ast)