Update revisionDesc
 - Submit - DCLP Workshop.
Please note the following: I am not sure the change of hands is displayed properly; I was not sure about how to deal with text added in the intercolumnium (col. ii, l. 3), so I encoded it as text added to the left of col. ii; no. of lines does not include those completely in lacuna. (Federica Micucci)
 - Vote - Accept-Straight-to-Finalization - In Display Name, change semicolon to comma; add CWKB numbers: 939 for the author, 2870 for the work; in the notes change to 'perhaps by the same second hand that also added the paragraphos.' In the text, encode the second col. as D=.ii.column; add modern punctuation to the Greek text. (Julia Lougovaya-Ast)
 - Finalized - Ready. (Julia Lougovaya-Ast)