Update revisionDesc
 - Submit - “The text was entered by participant of Dr. Julia Lougovaya’s Webinar on Greek Literary Ostraka, which was offered as part of Heidelberg’s "Webinars in Specialized Disciplines” initiative, sponsored by MWK-Baden Wuerttemberg” (Federica Micucci)
 - Vote - Return-Straight-to-Sender - Add under 'passage' line-spans for Homer ( Od. 11.311; 21.390-391); record West's reference for Antimachus, which fr. 57, rather than that in Powell, and make the formatting of the General Notes more homogeneous (ll., spacing around =, avoid ff.). In the text:
l. 1: change mark up at the end to [ca. 12]; l. 3, 7: record that the readings in the text are those of Esposito; l. 8: keep λινέοις of the ostracon and either delete the insertion of iota by 'edd.' or record the editors who insert it (West does not, so I would prefer simply to delete it). (Julia Lougovaya-Ast)
 - Submit - Changes made
The text was entered by participant of Dr. Julia Lougovaya’s Webinar on Greek Literary Ostraka, which was offered as part of Heidelberg’s "Webinars in Specialized Disciplines” initiative, sponsored by MWK-Baden Wuerttemberg (Federica Micucci)
 - Vote - Accept-Straight-to-Finalization - Looks good. (Julia Lougovaya-Ast)
 - Finalized - Ready. (Julia Lougovaya-Ast)