Update revisionDesc
 - Commit - TWO CORRECTIONS on the original readings IN l.. 2 and 9 (nikos litinas)
 - Submit - TWO CORRECTIONS on the original readings IN l.. 2 and 9
P.Corn. inv. MSS. A 101. XXII (= P.Fay. 315 description) was published by A. Calderini in P.Med. I, pp. 52-53 (= SB XVIII 13639). It is a receipt for syntaximon and its provenance is the village of Theadelphia in the Themistou meris of the Arsinoite nome. It is dated after April 25, A.D. 153 (BL IX, p. 304). The image can be found in http://quod.lib.umich.edu/a/apis/x-1280/c0_xxiir.tif. In l. 2 we read μετό]χ(οις) πράκ(τορσιν) Ἥρων Σαταβοῦτ̣[ος instead of ἀργυρικῶν Θεαδελφ(είας). For the payment of the syntaximon through the agents cf. P.Fay. 53, 4 (A.D. 111; Theadelphia); 54, 4 (A.D. 117-118; Euhemeria). For the tax syntaximon see P.Berl.Cohen 1, introd., p. 4.  In l. 9 after the month there is the day, so read Φαρμοῦθι κη̅ μετ̣[ὰ λόγον. For the meaning of the phrase μετὰ λόγον see P.Hamb. III 110, 7n. (nikos litinas)
 - Vote - Return-Straight-to-Sender - Sending back to let you look at this again, as we decided. (James Cowey)
 - Vote - Accept-Straight-to-Finalization - Fine (James Cowey)
 - Finalized - Ready. (James Cowey)