Update revisionDesc
 - Commit - l. 17: corrected accent from τὰ̓́λλα to τἆλλα; l. 19: altered "reg" to "corr", since obvious scribal error and not regularization (Thomas Backhuys)
 - Submit - l. 17: corrected accent from τὰ̓́λλα to τἆλλα; l. 19: altered "reg" to "corr", since obvious scribal error and not regularization (Thomas Backhuys)
 - Vote - AcceptText - I agree with the correction in l. 19.
"L. 17" should be "l. 16" in the comment. There is no consensus on whether ταλλα is the result of elision or crasis, so whether it should be spelt τἄλλα or τἆλλα. I don't have access to the printed edition at present, but I suspect the DDBDP text is the result of an encoding error for τἄλλα, in which case we should probably keep what the editor preferred. If it turns out the issue started with the printed edition, then I agree it should be corrected, and either τἄλλα or τἆλλα would be fine. (Myrto Malouta)
 - Vote - AcceptText - Encode <:τὰ ἄλλα|reg|τἆλλα:> to be consistent with papyri.info practice. Use simple pointed brackets in lin19 (James Cowey)
 - Finalized - Ready. (James Cowey)