Update revisionDesc
 - Submit - Added the English translation (from APIS) and added my own Arabic translation too. (Usama Gad)
 - Vote - AcceptWithModification - In the Arabic translation, in line 1. add أغسطس before جيرمانيكوس at the end of the line. In line 2. in stead of أوينيس write أونيس.
In The English translation, in line 3. instead Aulaios (translation of the old reading) write Aunes. (Usama Gad)
 - Vote - Accept - See Gad's comments.
And please also add after ديديموس in the Arabic translation to l.6 سكرتير الكاتب الملكي
and after Didymus in l.6 in the English translation, secretary of the royal scribe,
and in l.1 replace تبيريوس by تيبريوس, please. (Fatma E. Hamouda)
 - Finalized - Ready. (James Cowey)