Update revisionDesc
 - Submit - Originally submitted to litpap.info 2018.07.19 by Marco Perale. Submission has been transferred to papyri.info. Submission notes were: "I am in the process of entering all metrical pangrams. Deleted volume number in Editions & Bibliography/Preview; moved content from Legacy Display to Display; added info on script (both in Physical Description and General Notes); added MP3 number; added measurements; added bibliography" (James Cowey)
 - Vote - Accept-Straight-to-Finalization - Julia Lougovaya-Ast 2018-07-20: "Accept-Straight-to-Finalization - Looks good."
Marco Perale 2018-07-20 12:27:20: "Changes incorporated: see below. Text: note the likely omission in l. 6; YES, NOTED IN THE COMMENTARY in l. 8, is ψήφῳ a correction form ψήφι? In any case, why to postulate an iota adscr. as opposed to subscr.? YES, NOTED THE SCRIBAL CORRECTION IN LEIDEN+ In l. 13, there can hardly be more than one missing letter where [ ] is printed, or perhaps there is none at all? Can this be verified? IT’S EITHER NO SPACE OR ‘ι’; I THINK [.?] IS PREFERABLE OVER [.1?], AS IT IS TECHNICALLY CASE OF ‘Lost character gap unknown quantity’. In metadata: I'd suggest to record APl in parenthesis since a) it pertains to only one line of the text; b) the inclusion of that verse in APl is rather an evidence/attestation than a title. MAKES SENSE. CHANGED INTO: “alphabetic iambics and hexameters (incl. palindrome verse transmitted as Anth. Pal. 016 387c)”
Julia Lougovaya-Ast 2018-07-20: "Return-Straight-to-Sender - Text: note the likely omission in l. 6; in l. 8, is ψήφῳ a correction form ψήφι? In any case, why to postulate an iota adscr. as opposed to subscr.? In l. 13, there can hardly be more than one missing letter where [ ] is printed, or perhaps there is none at all? Can this be verified? In metadata: I'd suggest to record APl in parenthesis since a) it pertains to only one line of the text; b) the inclusion of that verse in APl is rather an evidence/attestation than a title." (James Cowey)
 - Finalized - Ready (Rodney Ast)