Update revisionDesc
 - Submit - Originally submitted to litpap.info 2018-03-21 by Anna Arpaia. Submission has been transferred to papyri.info. Submission note read: "I entered vac. in Col.II l.2 (PSI 11 1203 l.18), in order to reproduce that space left by the editor in lacuna: [.1 ] (actually only one letter is missing). I restored the missing parts of words in lacuna, even if in PSI 11 1203 we have only diplomatic transcription. On this regard, in col. III at the end of l. 27 (= PSI 1203, l. 54) I integrated τὰ ἄλ- (instead of τἄλ-), as in Col.II l.20 (=PSI 11 1203, l.36). Last: Col. III l.13 (=PSI 16 1605, l.13) I think των is a typing error, therefore I gave τον." (James Cowey)
 - Vote - Accept-Straight-to-Finalization - Julia Lougovaya-Ast 2018-11-15: "Accept-Straight-to-Finalization - Delete Legacy Name, adjust the language of the general note; in the text, note that Arpaia's PN changes are after a photo."
Anna Arpaia 2018-09-11: "Please, check col. II l. 1 (PSI XI 1203, col. II), where I read a line missing in the edition. I marked it as PN and wrote a note in the commentary, in order to explain my interpretation of the traces."
Julia Lougovaya-Ast 2018-06-14: "Return-Straight-to-Sender - I'd suggest to restore the text of col. 1 since the ends of the ends of the lines allow for determining the missing letters; I am not sure why ἀκοῦσαι in place of αἰτῆσαι is restored in col.i.5; there seems to be some text missing at the end of col. i in comparison with the MS tradition, could this be noted? Missing lines on top and bottom of the columns ought to be indicated less it look like the text runs continuously (which it does not); text in col. iii preserved in 1203 should be restored just as the text of that col. preserved in 1605 is given." (James Cowey)
 - Finalized - I was not convinced by restoration of col. 2, l. 1 and did not take it up; checked and corerected the text and added lines 39-41 of col. 3, which are visible on the photo but for some reason had been omitted from the edition in PSI 11.1203. (Julia Lougovaya)